lab_4/requirements.txt
filetype==1.2.0
python-dotenv==1.1.0
